+++ tests/acceptance/test_abort_commands.py
@@ -9,7 +9,6 @@ from ska_mid_dish_manager.models.dish_enums import (
     PointingState,
     TrackTableLoadMode,
 )
-from ska_mid_dish_manager.utils.ska_epoch_to_tai import get_current_tai_timestamp
 from tests.utils import az_el_slew_position
 
 
@@ -106,6 +105,7 @@ def test_abort_commands(
 def track_a_sample(
     event_store_class,
     dish_manager_proxy,
+    ds_device_proxy,
 ):
     """Execute a track command to slew the dish to a new position"""
 
@@ -155,7 +155,7 @@ def track_a_sample(
 
     # Load a track table
     current_az, current_el = dish_manager_proxy.achievedPointing[1:]
-    current_time_tai_s = get_current_tai_timestamp()
+    current_time_tai_s = ds_device_proxy.GetCurrentTAIOffset()
 
     # Directions to move values
     az_dir = 1 if current_az < 350 else -1